Testing with the Jasmine is set
Heroku tutorial #2

Demo

Jasmine getting started with #1

Jasmine-node is installed

Jasmine-node command npm test is set

Jasmine-node command npm test is defined for the project file

exampleSpec.js is created

Testing of server response is finally set

Testing of server response is correcting #1

Index.js is splitted into two modules: index.js+app.js. Testing is set #3

index.js is modified so the 'port' is defined by app.js

Negative test is set

Testing with Jasmine is set
