+++ app.js
@@ -1,0 +1,17 @@
+var express = require('express');
+var app = express();
+// var exports = module.exports = {};
+
+app.set('port', (process.env.PORT || 5000));
+
+app.use(express.static(__dirname + '/public'));
+
+// views is directory for all template files
+app.set('views', __dirname + '/views');
+app.set('view engine', 'ejs');
+
+app.get('/', function(request, response) {
+  response.render('pages/index');
+});
+
+module.exports = app;
