+++ index.js
@@ -1,20 +1,10 @@
-var express = require('express');
-var app = express();
 
-app.set('port', (process.env.PORT || 5000));
+var app = require('./app');
 
-app.use(express.static(__dirname + '/public'));
-
-// views is directory for all template files
-app.set('views', __dirname + '/views');
-app.set('view engine', 'ejs');
-
-app.get('/', function(request, response) {
-  response.render('pages/index');
-});
-
+// exports.closeServer = function(){
+//   console.log('Node app is closed on port', app.get('port'));
+//   server.close();
+// };
 app.listen(app.get('port'), function() {
   console.log('Node app is running on port', app.get('port'));
 });
-
-
