+++ package.json
@@ -4,9 +4,11 @@
   "description": "A sample Node.js app using Express 4",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js"
+    "start": "node index.js",
+    "test": "./node_modules/.bin/jasmine-node spec/exampleSpec.js"
   },
   "dependencies": {
+    "cool-ascii-faces": "1.3.3",
     "ejs": "2.3.3",
     "express": "4.13.3"
   },
@@ -22,5 +24,11 @@
     "heroku",
     "express"
   ],
-  "license": "MIT"
+  "license": "MIT",
+  "devDependencies": {
+    "cool-ascii-faces": "1.3.3",
+    "jasmine-node": "^1.14.5",
+    "request": "^2.69.0",
+    "supertest": "^1.2.0"
+  }
 }
